+++ database/migrations/2025_11_03_125844_create_cart_items_table.php
@@ -22,7 +22,7 @@ return new class extends Migration {
                 ->cascadeOnUpdate()
                 ->cascadeOnDelete();
 
-            $table->unsignedInteger('price');
+            $table->decimal('price');
             $table->unsignedInteger('quantity');
             $table->string('string_option_values')->nullable();
