+++ database/migrations/2025_11_17_084428_create_orders_table.php
@@ -28,7 +28,7 @@ return new class extends Migration {
                 ->index()
                 ->constrained('payment_methods');
 
-            $table->unsignedInteger('amount')->default(0);
+            $table->decimal('amount')->default(0);
             $table->timestamps();
         });
     }
